transformed from geolocation to sensor integration
