+++ homeassistant/components/geonetnz_volcano/const.py
@@ -3,16 +3,18 @@ from datetime import timedelta
 
 DOMAIN = "geonetnz_volcano"
 
-PLATFORMS = ("sensor", "geo_location")
-
 FEED = "feed"
 
-DEFAULT_ICON = "mdi:mountain"
+ATTR_ACTIVITY = "activity"
+ATTR_ALERT_LEVEL = "alert_level"
+ATTR_EXTERNAL_ID = "external_id"
+ATTR_HAZARDS = "hazards"
+
+# Icon alias "mdi:mountain" not working.
+DEFAULT_ICON = "mdi:image-filter-hdr"
 DEFAULT_RADIUS = 50.0
 DEFAULT_SCAN_INTERVAL = timedelta(minutes=5)
 
 SIGNAL_DELETE_ENTITY = "geonetnz_volcano_delete_{}"
+SIGNAL_NEW_SENSOR = "geonetnz_volcano_new_sensor_{}"
 SIGNAL_UPDATE_ENTITY = "geonetnz_volcano_update_{}"
-SIGNAL_STATUS = "geonetnz_volcano_status_{}"
-
-SIGNAL_NEW_GEOLOCATION = "geonetnz_volcano_new_geolocation_{}"
